# Add GGML_HIP_ROCWMMA_FATTN to enable rocWMMA for FlashAttention