… ggml/CMakeLists.txt
… ggml/src/ggml-cuda/fattn.cu
… ggml/src/ggml-hip/CMakeLists.txt