@JohannesGaessler 2025-02-22
As of right now I do not have the hardware necessary to test this code. Is anyone pledging to maintain it long-term if it gets merged?