@Headcrabed 2025-02-23
@JohannesGaessler So we will keep this code path and @adelj88 @thamwangjun 's optimized code path together in the future?