@Beinsezii 2025-02-25
Nice, +41% over master for pp2048 on gfx1100. Shame it doesn't affect text gen, where plain radv is still +22% over hip.