@Beinsezii 2025-02-25
@Headcrabed gfx12 is part of the current release 6.3
https://github.com/ROCm/rocWMMA/blob/develop/CHANGELOG.md#rocwmma-160-for-rocm-630