@bjj 2025-02-25
@hjc4869 `if (NOT ${FOUND_ROCWMMA})` would need to be `if (NOT FOUND_ROCWMMA)` to make that test work, because if you expand it, it ends up as a string, and the string is truthy.

However, `CHECK_INCLUDE_FILE_CXX` does not really work on `rocwmma/rocwmma.hpp` because it needs an include path (can add `"-I${ROCM_PATH}/include"` to the check) BUT ALSO other things included first (at least `stdint.h` and some other ROCm headers) so right now the patch relies on skipping the fatal error, even though cmake prints a line about not finding the header.