@JohannesGaessler 2025-02-25
>Let me know if there are any other tests I can run.

Check the output of `test-backend-ops -o FLASH_ATTN_EXT`.