@hjc4869 2025-02-25
With the `Q->ne[1] > 8` check removed, it's improving the perf of low batch size quite significantly.

`ROCR_VISIBLE_DEVICES=1 llama-batched-bench -ngl 999 -m llama3.1-8b-q4_0.gguf -fa -npl 1,2,4,6,8,10,12 -npp 512 -ntg 128 -c 16384`

Before:

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.204 |  2507.67 |    1.611 |    79.44 |    1.815 |   352.53 |
|   512 |    128 |    2 |   1280 |    0.361 |  2837.10 |    1.595 |   160.52 |    1.956 |   654.47 |
|   512 |    128 |    4 |   2560 |    0.748 |  2739.11 |    2.230 |   229.64 |    2.977 |   859.84 |
|   512 |    128 |    6 |   3840 |    1.174 |  2617.24 |    4.354 |   176.40 |    5.528 |   694.71 |
|   512 |    128 |    8 |   5120 |    1.636 |  2504.34 |    5.324 |   192.33 |    6.960 |   735.66 |
|   512 |    128 |   10 |   6400 |    2.135 |  2398.53 |    3.898 |   328.41 |    6.032 |  1060.98 |
|   512 |    128 |   12 |   7680 |    2.671 |  2300.17 |    4.162 |   369.08 |    6.833 |  1123.98 |

After
|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.202 |  2530.23 |    1.362 |    **93.99** |    1.564 |   409.14 |
|   512 |    128 |    2 |   1280 |    0.359 |  2852.48 |    1.521 |   **168.33** |    1.880 |   680.91 |
|   512 |    128 |    4 |   2560 |    0.748 |  2736.58 |    1.842 |   **277.90** |    2.591 |   988.11 |
|   512 |    128 |    6 |   3840 |    1.169 |  2627.11 |    2.377 |   **323.14** |    3.546 |  1082.91 |
|   512 |    128 |    8 |   5120 |    1.619 |  2530.44 |    2.800 |   **365.66** |    4.419 |  1158.60 |
|   512 |    128 |   10 |   6400 |    2.108 |  2429.33 |    3.847 |   332.72 |    5.955 |  1074.78 |
|   512 |    128 |   12 |   7680 |    2.636 |  2330.90 |    4.108 |   373.94 |    6.744 |  1138.86 |